Moving to src folder, fixing includes to compile
